Class Loader for Java block modified (compatibility for newer JVM's).
GeckoCIRCUITS assumed the system class loader to be a URLClassLoader, which it was
implicitely in older Java releases (e.g. 13).
However, this was an internal feature, and with some java release, they did change this
internal implementation, which made the Java-Block unusable. This bugfix should
solve the problem: GeckoBlockClassLoader itself was made an URLClassLoader.